rename NuGraph2 file to conform to snake case style
